backend/requirements_env.txt
# backend/requirements.txt

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
anthropic==0.7.8

# Image processing and OCR
pillow==10.1.0
pytesseract==0.3.10

# PDF processing
PyPDF2==3.0.1
pdf2image==1.16.3

# Audio processing
openai-whisper==20231117

# YouTube
youtube-transcript-api==0.6.1

# Utilities
python-dotenv==1.0.0
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0


# backend/.env.example

# Anthropic API Key (Required)
ANTHROPIC_API_KEY=your_api_key_here

# Server Configuration
HOST=0.0.0.0
PORT=8000
DEBUG=False

# File Upload Settings
MAX_FILE_SIZE_MB=10
ALLOWED_EXTENSIONS=jpg,jpeg,png,pdf,mp3,wav,m4a

# Logging
LOG_LEVEL=INFO
LOG_FILE=app.log

# CORS
CORS_ORIGINS=http://localhost:3000,http://localhost:3001

# Optional: OpenAI API Key (for Whisper alternative)
# OPENAI_API_KEY=your_openai_key_here